+++ src/services/planService.ts
@@ -4,6 +4,68 @@ import {isNullOrUndefined} from "../utils/helpers";
 import {PlanMember, ProjectMember} from "../models/Member";
 import {Plan} from "../models/Plan";
 import {Project} from "../models/Project";
+import {
+  ProjectCollaboratorAccessLevel
+} from "../models/Collaborator";
+import {isAdmin, isSuperAdmin} from "./authService";
+import {User} from "../models/User";
+import {prepareObjectForLogs} from "../logger";
+
+export const hasPermissionOnPlan = async (
+  context: MyContext,
+  plan: Plan,
+  requiredAccessLevel = ProjectCollaboratorAccessLevel.EDIT,
+): Promise<boolean> => {
+  const reference = 'planService.hasPermissionOnPlan';
+  if (!context || !context.token) return false;
+
+  // Super admins always have permission
+  if (await isSuperAdmin(context.token)) {
+    return true;
+  }
+
+  if (plan && plan.id) {
+    // See if the plan is listed in the token
+    const tokenDMP = context.token.dmpIds.find((entry) => {
+      return entry.dmpId === plan.dmpId
+    });
+
+    // If the user created the plan then they have access
+    if (plan.createdById === context.token.id) {
+      return true;
+    }
+
+    // If the current user is an Admin and the creator of the plan has the same affiliation
+    if (await isAdmin(context.token)) {
+      const planCreator = await User.findById(reference, context, plan.createdById);
+      if (planCreator && planCreator.affiliationId === context.token.affiliationId) {
+        return true;
+      }
+    }
+
+    // The DMP was not listed on their token then they do not have access
+    if (!tokenDMP) {
+      return false;
+    }
+
+    // Otherwise check to make sure the user has the desired access level
+    switch (requiredAccessLevel) {
+      case ProjectCollaboratorAccessLevel.COMMENT:
+        return true;
+      case ProjectCollaboratorAccessLevel.EDIT:
+        return tokenDMP.accessLevel === ProjectCollaboratorAccessLevel.OWN ||
+          tokenDMP.accessLevel === ProjectCollaboratorAccessLevel.EDIT;
+      case ProjectCollaboratorAccessLevel.OWN:
+        return tokenDMP.accessLevel === ProjectCollaboratorAccessLevel.OWN;
+      default:
+        return false;
+    }
+  }
+
+  const payload = { planId: plan?.id, userId: context.token?.id };
+  context.logger.error(prepareObjectForLogs(payload), `AUTH failure: ${reference}`);
+  return false;
+}
 
 export async function updateMemberRoles(
   reference: string,
@@ -93,3 +155,4 @@ export const ensureDefaultPlanContact = async (
   }
   return false
 }
+